Fixed GCC warnings (different signedness comparison) (#1739)
* Fix issue #1738
* Remove index >= 0 since we do unnecessary comparison

+++ synfig-studio/src/gui/states/state_bone.cpp
@@ -1089,7 +1089,7 @@ StateBone_Context::find_bone(Point point,Layer::Handle layer,int lay)const
 		Real close_line(10000000),close_origin(10000000);
 		Vector direction;
 		Angle angle;
-		int ret;
+		unsigned long ret;
 		Matrix m;
 		for(auto iter=list.begin();iter!=list.end();++iter){
